+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Added
 * Added `audioStreamType` in `AudioVideoConfiguration` for supporting audio stream configuration.
-
+* Added support for background blur and replacement video filter processor.
+* [Demo] Adder two new video filters - background blur and replacement.
 ### Fixed
 * Fixed calling start multiple times without stop crashes video in some phones. (Issue #356)